New method for Innate::Session, #resid!, that can be used to change the session key.
The intent of the method is to allow changing session keys when user logs in
to prevent certain session fixation attack vectors. #resid! moves all
session data under a new session key and thus invalidates the old key.

It is probably a good idea to explicitly reset the session data as well on
login, but the necessity of this depends on the application.